add a pillar option to skip salt configuration
Partially fixes https://github.com/sans-dfir/sift/issues/272 as
the salt minion will not be stopped if sift_manage_salt is set
to False.

This option does not change the default behavior of the salt
state that stop and disable salt-minion service.

A proper documentation of all pillar options that can change the
behavior of the sift salt states has also been created.